* changed return value into double in ProfileData * refactored remaining code

+++ app/src/main/java/pl/projektorion/krzysztof/blesensortag/bluetooth/SensorTag/ProfileData.java
@@ -5,6 +5,6 @@ package pl.projektorion.krzysztof.blesensortag.bluetooth.SensorTag;
  */
 
 public interface ProfileData {
-    int getValue(int sensorAttribute);
+    double getValue(int sensorAttribute);
     byte[] getRawData();
 }
